Add p tags
Text should nearly always be wrapped in a p (paragraph) tag, unless it's a h1 or other kind of text. using double <br>s are generally frowned upon as "visual" not semantic HTML, as you're describing the how it looks, not what the text means.

I've left one br in, as you're using it to force a linebreak, which is a better use of it (although not perfect, but perfectly acceptable)