Add main tag
Adding widths, margins and paddings to the body can behave a bit weirdly, so best to wrap everything in its own tag within the body. This could have been a div, but the slightly more semantic thing to do these days is use the trendy new main tag.

"width: 80%" wasn't really doing anything, so I've removed it.
I've added a "max-width", which means that the page won't get any wider than 750px, so it'll look nicer on big screens. This means I can get rid of the <br> tag, which means the text also wraps nicer on small screens.